requirements.txt
groq
pandas
plotly
streamlit
python-dotenv
